# backend/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Visibility": {
				"enum": [
					"public",
					"logged_in",
					"private"
				],
				"type": "string"
			},
			"Item": {
				"properties": {
					"id": {
						"type": "string"
					},
					"visibility": {
						"$ref": "#/components/schemas/Visibility"
					},
					"source": {
						"type": "string",
						"enum": [
							"5th_e_SRD",
							"Homebrew"
						],
						"nullable": false
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "number",
						"format": "double"
					},
					"imageId": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"shortDescription": {
						"type": "string"
					},
					"mainDescription": {
						"type": "string"
					},
					"price": {
						"type": "string",
						"nullable": true
					},
					"rarity": {
						"type": "string",
						"nullable": true
					},
					"weight": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"features": {
						"items": {
							"properties": {
								"featureDescription": {
									"type": "string"
								},
								"featureName": {
									"type": "string"
								}
							},
							"required": [
								"featureDescription",
								"featureName"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"visibility",
					"source",
					"createdBy",
					"createdAt",
					"imageId",
					"name",
					"shortDescription",
					"mainDescription",
					"price",
					"rarity",
					"weight",
					"features"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ImageMetadata": {
				"properties": {
					"id": {
						"type": "string"
					},
					"visibility": {
						"$ref": "#/components/schemas/Visibility"
					},
					"source": {
						"type": "string",
						"enum": [
							"5th_e_SRD",
							"Homebrew"
						],
						"nullable": false
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "number",
						"format": "double"
					},
					"fileName": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"ownerId": {
						"type": "string"
					},
					"ownerType": {
						"type": "string",
						"enum": [
							"item",
							"weapon",
							"monster"
						],
						"nullable": false
					}
				},
				"required": [
					"id",
					"visibility",
					"source",
					"createdBy",
					"createdAt",
					"fileName",
					"mimeType",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Image": {
				"properties": {
					"metadata": {
						"$ref": "#/components/schemas/ImageMetadata"
					},
					"base64": {
						"type": "string"
					}
				},
				"required": [
					"metadata",
					"base64"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemUpdateResponse": {
				"properties": {
					"item": {
						"$ref": "#/components/schemas/Item"
					},
					"image": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Image"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"item"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemUpdateRequest": {
				"properties": {
					"item": {
						"$ref": "#/components/schemas/Item"
					},
					"image": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Image"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"item"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/items/{userId}": {
			"get": {
				"operationId": "GetAllForUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Item"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/item": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Item"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": []
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemUpdateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ItemUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/item/{itemId}": {
			"get": {
				"operationId": "GetItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Item"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteIem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Item"
								}
							}
						}
					}
				},
				"tags": [
					"Item"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}